Adds propidl.h
This adds most of the contents of propidl.h to comtypes. I have found that there is a lot of COM interfaces that use the IPropertyStore interface along with the PROPVARIANT structure. I thought it would be nice to include this in comtypes like the VARIANT structure has been added.