Fixes tools.codegenerator.Generator.generate_code
I ran across an issue where I had exported a typelib resource from a dll file and the resulting tlb file when fed into the code generator would fail because of a version mismatch. This was happening because the file stored in the typelib was different from the file that was being used to generate the code.