fix(npsupport): make numpy typecodes consistent across numpy versions (#239)
In newer versions of numpy the ctypeslib._typecodes variable is no longer
present, so comtypes populates it itself if it is not found. Previously the
created dictionary was a mapping of ctypes classes to numpy dtypes, whereas
the previous _typecodes dictionary mapped numpy dtype string representations
to ctypes classes. This fix rearranges the mapping in the dictionary to be
consistent with the previous format.

Closes #238

Co-authored-by: Ben Rowland <ben.rowland@hallmarq.net>