Fixing the tests
This is the start of getting the tests up and running again. Quite a few of them are running properly and they are passing. I have an error in the test_jscript.js file I am not sure where that's coming from. Maybe someone else could shed some light on that bit of the code for me.

 At this point the above error is not my major concern. The tests don't finish and I get an application crash 0xC0000005 (Access Violation). I have to look into what is going on with this.

 I did fix compiling the typelib in test_server.py.

 Over the last 2 years or so I have been working on a script that creates an MSVC build environment for compiling Python extension modules and C libraries. The one that both distutils and setuptools creates is riddled with problems and neither of them work on all MSVC versions. The script I wrote DOES NOT use the vcvars*.bat files. to create the build environment. Those batch files are buggy and don't always setup the build environment properly.

It's kind of circular because the script uses comtypes to create the build environment. While it can do it without as a backup it is much faster with it. So even if comtypes  is broken in a manner that does not allow that portion of the script to run the server test type lib will still get compiled properly.

The MSVC script works for Visual Studio 2008 up to the most current Visual Studio 2022, Visual C++ 2008 to current Visual C++ 2022 All versions of  Build Tools. It supports all .NET and NETFX versions and all Windows SDK's from 6.0a to the latest Windows 11 SDK.

It should work without issue for many years.

I forgot, it detects the Visual C++ Compiler for Python 2.7 as well.