Cleans up generated code output.
This is a pretty involved PR in what it does. There is a decent amount of code changes but nothing to crazy. Here are the things the PR does

removes duplicate imports
removes unused imports
puts declarations after imports instead of being mixed in with the imports
formats most code that is over 80 characters on a single line
2 blank line before and after classes
puts class docstrings as the first line in the class  this way it can actually get set to a docstring
fixes lack of triple double quotes not being used around class docstrings

So in a nutshell it makes most of the generated code PEP8 compliant and it makes the generated code a whole hell of a lot easier on the :eyes: