drop some support for legacy numpy <1.6
This commit just removes some rather hacky code designed to allow comtypes to
work with numpy's datetime64 class at a time when it might not exist in older
versions of numpy. However, now that we are some 17 versions of numpy further
on, we can probably drop support for these legacy versions.

This commit also modified the way that numpy is imported in test_npsupport.py
as previously the check to see if numpy supported datetime was being run
before the check to see if numpy was importable, so the check always failed and
datetime related tests were not being run at all, even when supported