improve namespacing of numpy interop
This commit moves the existing npsupport.py to _npsupport.py, and adds a line
in __init__.py to import the interop object from _npsupport as npsupport in
the main comtypes namespace. This means that the original functions like
"isndarray()", "isdatetime64()" etc. are still available at
comtypes.npsupport.isndarray() etc. as well as comtypes.npsupport.enable(),
rather than requiring comtypes.npsupport.interop.enable() as previously.

In addition, it adds some features to the test suite - in particular there is a
custom decorator which replaces the decorated test method with two new ones,
one with npsupport enabled and the other with it disabled, testing that the
former passes the test, and the latter raises the correct error.