333 optin numpy v2 (#343)
* feat: make numpy interop opt-in

This commit removes all automatic imports of numpy, and instead adds a new
function "enable_numpy_interop()" to comtypes.npsupport. This function is
the only one which actually imports numpy, any attempt to use numpy interop
related features without calling "enable_numpy_interop()" first will lead to
an ImportError being raised with a message explaining that the function needs
to be called before using numpy functionality.

Other parts of comtypes wishing to access numpy functions should call
comtypes.npsupport.get_numpy() which will return the module.

* make npsupport.isndarray raise an Exception if interop not enabled

Without numpy interop being enabled, we can't directly check if a variable is
an ndarray, but having the __array_interface__ attribute is a fairly good
estimator, so if an object with that attribute is passed to isndarray before
interop is enabled, a ValueError will be raised prompting the user to call
npsupport.enable_numpy_interop().

* make safearray_as_ndarray automatically enable np interop

Entering the safearray_as_ndarray context manager will internally call
npsupport.enable_numpy_interop(), as it is clear that the user wants to have
numpy support.

early return from repeated calls to npsupport.enable_numpy_interop()

Calling enable_numpy_interop() when interop is already enabled returns at the
top of the function.

* reorganise tests relating to np interop

I have gathered all the numpy related tests into a single test file,
test_npsupport.py, which has a couple of different TestCases internally,
reflecting the original organisation. test_npsupport will only be run if
importing numpy succeeds.

I also removed a lot of @skip decorators relating to numpy dependence, all
tests currently pass (or are skipped) with or without numpy being installed
(on my system at least, Python 3.9.13 and numpy 1.23.1).

* fix syntax errors for older Python versions

Remove inline type annotations on a couple of functions in test_npsupport.py
and an f-string in npsupport.py

* refactor to use a singleton class instead of global variables

Modify npsupport to put all functionality inside a class Interop, which
exposes public interface methods "enable()", "isndarray()", "isdatetime64" and
the properties "numpy", "VARIANT_dtype", "typecodes", "datetime64" and
"com_null_date64". A singleton instance of the class (called interop) is
created in the npsupport namespace, to use numpy interop its "enable()" method
should be called. It is also still valid to use the "safearray_as_ndarray"
context manager to enable support as well.

* fix bug with numpy.ctypeslib._typecodes member

The previous version introduced a bug by globally returning a variable that is
only created for newer versions of numpy, this commit reverts that single line
change.

* drop some support for legacy numpy <1.6

This commit just removes some rather hacky code designed to allow comtypes to
work with numpy's datetime64 class at a time when it might not exist in older
versions of numpy. However, now that we are some 17 versions of numpy further
on, we can probably drop support for these legacy versions.

This commit also modified the way that numpy is imported in test_npsupport.py
as previously the check to see if numpy supported datetime was being run
before the check to see if numpy was importable, so the check always failed and
datetime related tests were not being run at all, even when supported

* improve namespacing of numpy interop

This commit moves the existing npsupport.py to _npsupport.py, and adds a line
in __init__.py to import the interop object from _npsupport as npsupport in
the main comtypes namespace. This means that the original functions like
"isndarray()", "isdatetime64()" etc. are still available at
comtypes.npsupport.isndarray() etc. as well as comtypes.npsupport.enable(),
rather than requiring comtypes.npsupport.interop.enable() as previously.

In addition, it adds some features to the test suite - in particular there is a
custom decorator which replaces the decorated test method with two new ones,
one with npsupport enabled and the other with it disabled, testing that the
former passes the test, and the latter raises the correct error.

Co-authored-by: Ben Rowland <ben.rowland@hallmarq.net>