Rebase on main.
Added an interface to the C++ COM server for testing pointers to safearrays
as method parameres of a Dispatch Interface.

The implementation of the interface cannot share the IDispatch implementation
with other interfaces and requires a separate component.
(see https://learn.microsoft.com/en-us/cpp/atl/multiple-dual-interfaces)

Therefore the naming scheme of the existing component that implements the
test interface for record pointers was also changed.