Changed the name of the component that implements the interface for record parameter testing.
Background: Implementing several related interfaces in a single
component is a problem when the interfaces inherit from the same
interface but require different implementations for this interface
as is the case for IDispatch. Therefore additional interfaces will
be implemented in separate components.

The name change reflects the limited scope of the component.