Changed the default value of the 'extra' parameter in the 'create' method of tagSAFEARRAY subtypes which are generated by '_midlSAFEARRAY' to the setting for 'extra' identified in the '_make_safearray_type' method.
The 'extra' value required for the creation of SAFEARRAYs with VT_RECORD,
VT_DISPATCH or VT_UNKNOWN elements is already identified at the time when
the associated tagSAFEARRAY subtype is created in the '_make_safearray_type'
method. Using this value by default for the 'extra' parameter when creating
an instance of the type does avoid performing the same identification process
again in user code.