Add `test_typeannotator`.
Code within `TYPE_CHECKING` blocks are not executed at runtime, so they will
not result in `NameError` or `AttributeError`.
However, if method names or argument names use Python reserved words, or if
the statements or expressions are incorrect in Python syntax, a `SyntaxError`
will be raised.

The newly added tests ensure that `typeannotator` avoids such situations:
- If there are method names that match Python reserved words,
  `typeannotator` replaces them with `pass` and comments.
- If there are argument names that match Python reserved words,
  `typeannotator` replaces them with `*args` or `**kwargs`.
- If there are methods that specify required arguments after optional
  arguments, `typeannotator` replaces them with `**kwargs`.

It was challenging to find a "real-world" COM library that covered all these
cases. Therefore, in these tests, using `typedesc`s as a test double to
simulate the results of parsing a COM type library.

It has been confirmed through the discussions in the following issue that
the changes made to the production code in commit 35f3336 prevent errors from
occurring in a real COM library.
https://github.com/enthought/comtypes/issues/588#issuecomment-2230209573