Support Python 3.14 and drop Python 3.8. (#839)
* Add a version bridge to `_post_coinit/unknwn.py`
to accommodate the changes in `ctypes` in Python 3.14, including the
deprecation of `_pointer_type_cache` and the introduction of the
`__pointer_type__` attribute protocol.

* Add a version bridge to `_meta.py`
to accommodate the changes in `ctypes` in Python 3.14, including the
deprecation of `_pointer_type_cache` and the introduction of the
`__pointer_type__` attribute protocol.

* Improve error message for `PyCArgObject` size check.

* Update PyCArgObject.value for Python 3.14 compatibility.

* Set packing for `PyCArgObject` on Python 3.14+.

In Python 3.14, changes to the underlying C structures in `ctypes`
require more explicit control over memory layout. This change
enforces 8-byte alignment, which correctly pads the structure
and ensures that the `value` field is accessed at the correct
offset.
This is critical for preventing memory corruption and maintaining
compatibility on both 32-bit and 64-bit systems.

* Update the autotest GHA workflow.

* Remove `allow-prereleases: true` from `autotest.yml`.

* Drop Python 3.8 support.

* Add the version bridge for `test_comserver.TestInproc_win32com.test_eval`.